wip: fixing up mobile
