+++ src/components/user_info/components/chart/components/data_diagram/index.tsx
