+++ src/components/user_info/components/chart/index.test.tsx
