+++ src/index.ts
@@ -33,7 +33,7 @@ import ValidatorProfile from './components/validator_profile';
 import ProposalResultDesktop from './components/proposal_result_desktop';
 import ProposalResultMobile from './components/proposal_result_mobile';
 import LatestActivities from './components/latest_activities';
-import UserInfoMobile from './components/user_info_mobile';
+import UserInfo from './components/user_info';
 import UserInfoDesktop from './components/user_info_desktop';
 import InfoDialog from './components/info_dialog';
 
@@ -73,7 +73,7 @@ export {
   ProposalResultDesktop,
   ProposalResultMobile,
   LatestActivities,
-  UserInfoMobile,
+  UserInfo,
   UserInfoDesktop,
   InfoDialog,
 };
